requirements.txt
openai
ddtrace
python-dotenv